fix(JwtConfigProperties): expireLength의 타입을 String으로 변경

+++ src/main/kotlin/nexters/linkllet/common/support/JwtProvider.kt
@@ -24,7 +24,7 @@ class JwtProvider(
         return Jwts.builder()
             .setClaims(claims)
             .setIssuedAt(now)
-            .setExpiration(Date(now.time + jwtConfigProperties.expireLength))
+            .setExpiration(Date(now.time + jwtConfigProperties.expireLength.toLong()))
             .signWith(signingKey)
             .compact()
     }
